requirements.txt
# Core dependencies
python-binance>=1.0.17
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
python-dateutil>=2.9.0

# Data visualization
plotly>=5.14.0

# HTTP requests
requests>=2.31.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Future: AI components (to be added later)
# anthropic>=0.8.0
# langchain>=0.1.0
# crewai>=0.1.0
